📦 NEW: add Alfred snippet collection
A snippet collection for Alfred PowerPack users.

Intentionally, the "Name" column of the collection does not start with
an emojji. Alfreds Snippet browser doesn't use fuzzy find, it appears
to work off of first letters. Having the emojjis as the lead chatacter
makes the words hard to browse this way.

With auto expand turned on, the following keywords will expand to:

;new = 📦 NEW:
;improve = 👌 IMPROVE:
;fix = 🐛 FIX:
;release = 🚀 RELEASE:
;doc = 📖 DOC:
;test = ✅ TEST: