📦 NEW: Alfred snippet collection (#14)
* 📦 NEW: add Alfred snippet collection

A snippet collection for Alfred PowerPack users.

Intentionally, the "Name" column of the collection does not start with
an emojji. Alfreds Snippet browser doesn't use fuzzy find, it appears
to work off of first letters. Having the emojjis as the lead chatacter
makes the words hard to browse this way.

With auto expand turned on, the following keywords will expand to:

;new = 📦 NEW:
;improve = 👌 IMPROVE:
;fix = 🐛 FIX:
;release = 🚀 RELEASE:
;doc = 📖 DOC:
;test = ✅ TEST:

* 👌 IMPROVE: use "g" namespace w/ shorter keywords

It's easier to remember and namespaced snippets since g stands for git.

Shorter keywors match shell function naming.

* 🐛 FIX: format TODO properly for todo bot

the bot expects the flag at the begining of a line, not the end.

* 📦 NEW: Icon